+++ lexicmap/cmd/lib-seq_compare.go
@@ -434,7 +434,7 @@ func (cpr *SeqComparator) Compare(begin, end uint32, s []byte, queryLen int) (*S
 	// }
 	// fmt.Println("-------------------------------")
 
-	TrimSubStrPairs(poolSub2, subs, k)
+	TrimSubStrPairs(poolSub2, subs, k, 100)
 	if len(*subs) == 0 {
 		RecycleSubstrPairs(poolSub2, subs)
 		return nil, nil
@@ -526,7 +526,7 @@ func (cpr *SeqComparator) RecycleIndex() {
 //	727: 789-819 (-) vs 789-819 (-), len:31
 //	728: 790-820 (-) vs 790-820 (-), len:31
 //	729: 804-821 (-) vs 821-838 (-), len:18 <--- gap=17, overlap=17 (17/18)
-func TrimSubStrPairs(poolSub *sync.Pool, subs *[]*SubstrPair, k int) {
+func TrimSubStrPairs(poolSub *sync.Pool, subs *[]*SubstrPair, k int, minDist float64) {
 	if len(*subs) < 2 {
 		return
 	}
@@ -540,8 +540,9 @@ func TrimSubStrPairs(poolSub *sync.Pool, subs *[]*SubstrPair, k int) {
 	start := 0
 
 	for i, p = range (*subs)[1:] {
-		if (p.QBegin == _p.QBegin || p.TBegin == _p.TBegin) || // case 1
-			(gap(_p, p) > 11 && float64(overlap(_p, p))/float64(_p.Len) > 0.8) { // case 2
+		if distance(p, _p) < minDist &&
+			((p.QBegin == _p.QBegin || p.TBegin == _p.TBegin) || // case 1
+				(gap(_p, p) > 11 && float64(overlap(_p, p))/float64(_p.Len) > 0.8)) { // case 2
 			start = i
 			_p = p
 			continue
@@ -557,8 +558,9 @@ func TrimSubStrPairs(poolSub *sync.Pool, subs *[]*SubstrPair, k int) {
 	for i = len(*subs) - 2; i >= 0; i-- {
 		p = (*subs)[i]
 
-		if (p.QBegin == _p.QBegin || p.TBegin == _p.TBegin) || // case 1
-			(gap(p, _p) > 11 && float64(overlap(p, _p))/float64(_p.Len) > 0.8) { // case 2
+		if distance(p, _p) < minDist &&
+			((p.QBegin == _p.QBegin || p.TBegin == _p.TBegin) || // case 1
+				(gap(p, _p) > 11 && float64(overlap(p, _p))/float64(_p.Len) > 0.8 && distance(p, _p) < minDist)) { // case 2
 			end = i
 			_p = p
 			continue